initial commit in DB branch i have developed Departmentstore still have work to do

+++ backend/models/models.go
@@ -18,33 +18,33 @@ type User struct {
 
 // Faculty corresponds to the FACULTIES table.
 type Faculty struct {
-	FacultyID int    `json:"faculty_id"`
+	FacultyID int64  `json:"faculty_id"`
 	Name      string `json:"name"`
 }
 
 // Department corresponds to the DEPARTMENTS table.
 type Department struct {
-	DepartmentID   int    `json:"department_id"`
+	DepartmentID   int64  `json:"department_id"`
 	DepartmentName string `json:"department_name"`
-	FacultyID      int    `json:"faculty_id"`
+	FacultyID      int64  `json:"faculty_id"`
 }
 
 // Instructor corresponds to the INSTRUCTORS table.
 type Instructor struct {
-	InstructorID int    `json:"instructor_id"`
-	UserID       int    `json:"user_id"`
+	InstructorID int64  `json:"instructor_id"`
+	UserID       int64  `json:"user_id"`
 	FirstName    string `json:"first_name"`
 	LastName     string `json:"last_name"`
-	DepartmentID *int   `json:"department_id,omitempty"` // Use a pointer for nullable foreign keys.
+	DepartmentID *int64 `json:"department_id,omitempty"` // Use a pointer for nullable foreign keys.
 }
 
 // Student corresponds to the STUDENTS table.
 // fixx add national_id to be string
 type Student struct {
-	StudentID    int        `json:"student_id"`
-	UserID       int        `json:"user_id"`
-	Cgpa         float32   `json:"cgpa,omitempty"`
-	DepartmentID *int       `json:"department_id,omitempty"`
+	StudentID    int64      `json:"student_id"`
+	UserID       int64      `json:"user_id"`
+	Cgpa         float32    `json:"cgpa,omitempty"`
+	DepartmentID *int64     `json:"department_id,omitempty"`
 	FirstName    string     `json:"first_name"`
 	LastName     string     `json:"last_name"`
 	DateOfBirth  *time.Time `json:"date_of_birth,omitempty"` // Use a pointer for nullable dates.
@@ -52,12 +52,12 @@ type Student struct {
 
 // Course corresponds to the COURSES table.
 type Course struct {
-	CourseID     int     `json:"course_id"`
+	CourseID     int64   `json:"course_id"`
 	CourseCode   string  `json:"course_code"`
 	Title        string  `json:"title"`
 	Description  *string `json:"description,omitempty"` // Use a pointer for nullable text fields.
-	Credits      int     `json:"credits"`
-	DepartmentID int     `json:"department_id"`
+	Credits      int64   `json:"credits"`
+	DepartmentID int64   `json:"department_id"`
 }
 
 // CoursePrerequisite corresponds to the COURSE_PREREQUISITES table.
